feat(SemanticCheck.h/cc): 在SymbolEntry中加入can_calculated字段，完善canCalculated中对于Lval类型的处理

+++ semantic/SemanticCheck.h
@@ -60,7 +60,7 @@ private:
         return curr_while_depth_ != 0;
     }
 
-    bool canCalculated(Expression *init, float *value);
+    bool canCalculated(Expression *init, double *value);
 
     void addLibFunc();
